data: fix duplicated entries, use binary representation
